+++ relay-dynamic-config/src/feature.rs
@@ -22,6 +22,9 @@ pub enum Feature {
     /// Extract spans from transactions and convert them to standalone spans.
     #[serde(rename = "projects:extract-standalone-spans")]
     ExtractStandaloneSpans,
+    /// Allow ingestion of metrics in the "custom" namespace.
+    #[serde(rename = "organizations:custom-metrics")]
+    CustomMetrics,
 
     /// Deprecated, still forwarded for older downstream Relays.
     #[serde(rename = "organizations:transaction-name-mark-scrubbed-as-sanitized")]